Add multi-failure check before removing register
Adding a flag to be able to specify a number of failed checks before
mesos-consul removes the task from consul.

The motivation behind this change is because we had observed that
mesos-consul was removing running tasks from consul when the leader
mesos-master was down during the poll for running tasks. This would lead
to our load balancer getting reconfigured to have no services for a
significant period of time. The idea behind this solution is to
configure the number of times mesos-consul should tolerate not finding
the registered task in the mesos-master leader state before
de-registering it from Consul.