Replace _ from mesos task name by -
Previously underscores where removed when creating the service name.
This usually does not correspond to the intent of the user since he set
a separator in the task name.

In the specific case of marathon, the folder separator is converted to
underscore when creating a mesos task.

Users in marathon sets: /mygroup/mysubgroup/myapp

Tasks name is mesos are: mygroup_mysubgroup_myapp.[guid]

Before this patch, consul service names were mygroupmysubgroupmyapp
After this patch, consul service names are mygroup-mysubgroup-myapp.

Fixes #63